Support Dark Mode
Change from Firefox non-standard to standard system color values.

The options page needs to be opened in a tab as the Firefox UI and
website color schemes can diverge. Without open_in_tab, the options
page uses the website scheme and is embedded in a page using the Firefox
UI scheme. This can lead to clashing colors.